+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ Changelog
 3.4 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- Seating algorithm: improve performances by a factor of 4
 
 
 3.3 (2023-11-21)